requirements.txt
wheel
tqdm
pytorch_lightning==2.2.3
black
ruff
hydra-core==1.3.2
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib==3.8.4
moviepy==1.0.3
imageio
imageio-ffmpeg
# git+https://github.com/dcharatan/diff-gaussian-rasterization-modified
# git+https://github.com/HeliosZhao/language-gs-rasterization
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
opencv-python
sk-video

opencv-python
# diffusers==0.27.2
diffusers==0.30.3
accelerate==0.31.0
tensorboard
transformers
ipdb
easydict
# git+https://github.com/camenduru/simple-knn

pandas
# deepspeed==0.13.1
gpustat

decord
bitsandbytes
open3d

mediapy
einshape
mathutils


absl_py
gin_config

opencv_contrib_python
opencv_python
trimesh
pymeshlab
xatlas
plyfile
rawpy==0.20.0
ninja
scipy
scikit-image
scikit-learn
tensorboard
tensorboardX
